backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-openai
langchain-community
pypdf
youtube-transcript-api
faiss-cpu
python-dotenv
tiktoken
